Updated to read environment variables + renamed climberRouter
