+++ .gitignore
@@ -1,7 +1,9 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+node_modules
+FE/node_modules
+server/node_modules
 /.pnp
 .pnp.js
 
@@ -21,3 +23,8 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# ignore local environment variables,
+# everyone should make their own version
+# of this file
+.env
