+++ FE/src/components/EditClimber.js
@@ -10,7 +10,7 @@ const EditClimber = () => {
  
     const updateClimber = async (e) => {
         e.preventDefault();
-        await axios.patch(`http://localhost:5000/climbers/${id}`,{
+        await axios.patch(`http://localhost:4000/climbers/${id}`,{
             name: name
         });
         history("/");
@@ -21,7 +21,7 @@ const EditClimber = () => {
     }, []);
  
     const getClimberById = async () => {
-        const response = await axios.get(`http://localhost:5000/climbers/${id}`);
+        const response = await axios.get(`http://localhost:4000/climbers/${id}`);
         setName(response.data.name);
     }
