+++ server/config/database.js
@@ -1,9 +1,10 @@
 import { Sequelize } from "sequelize";
+import 'dotenv/config.js';
  
-const db = new Sequelize('testapp', 'root', 'helloworld', {
-    host: "localhost",
+const db = new Sequelize(`${process.env.DATABASE_NAME}`, `${process.env.SQL_USERNAME}`, `${process.env.SQL_PASSWORD}`, {
+    host: `${process.env.HOSTNAME}`,
     dialect: "mysql",
-    port: "3308"
+    port: `${process.env.SQL_PORT}`
 });
  
 export default db;
