+++ server/index.js
@@ -1,6 +1,7 @@
 import express from "express";
+import 'dotenv/config.js';
 import db from "./config/database.js";
-import climberRoutes from "./routes/index.js";
+import climberRoutes from "./routes/climberRouter.js";
 import cors from "cors";
 
 const app = express();
@@ -16,4 +17,4 @@ app.use(cors());
 app.use(express.json());
 app.use('/climbers', climberRoutes);
  
-app.listen(5000, () => console.log('Server running at port 5000'));
+app.listen(process.env.SERVER_PORT, () => console.log(`Server running at port ${process.env.SERVER_PORT}`));
