+++ server/models/ClimberModel.js
@@ -1,4 +1,5 @@
 import { Sequelize } from "sequelize";
+import 'dotenv/config.js';
 import db from "../config/database.js";
  
 const { DataTypes } = Sequelize;
@@ -17,7 +18,7 @@ const Product = db.define('Climber',{
     //     type: DataTypes.STRING
     // }
 },{
-    freezeTableName: true
+    tableName: `${process.env.CLIMBER_TABLE}`
 });
  
 export default Product;
